Fixed our team css and moved components into directory
