+++ frontend/src/app/team/page.module.css
@@ -1,0 +1,39 @@
+@import url("https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Roboto+Slab:wght@100..900&display=swap");
+
+.description {
+  text-align: left;
+  font-family: "Open Sans";
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 32px; /* 133.333% */
+}
+
+.subtitle {
+  font: var(--font-small-subtitle);
+  font-size: 32px;
+  line-height: 150%; /* 48px */
+  letter-spacing: 0.64px;
+}
+
+.text {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 47px;
+  margin-left: 99px;
+  margin-right: 99px;
+  margin-top: 53px;
+  margin-bottom: 64px;
+}
+
+.membersContainer {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  margin-left: 119px;
+  margin-right: 119px;
+  margin-bottom: 147px;
+  gap: 60px;
+}
