+++ frontend/src/app/team/page.tsx
@@ -2,10 +2,11 @@
 
 import React, { useEffect, useState } from "react";
 
-import BackgroundHeader from "./BackgroundHeader";
-import MemberInfo from "./MemberInfo";
+import styles from "./page.module.css";
 
 import { Member, getAllMembers } from "@/api/member";
+import BackgroundHeader from "@/components/BackgroundHeader";
+import MemberInfo from "@/components/MemberInfo";
 
 export default function Team() {
   const [members, setMembers] = useState<Member[]>([]);
@@ -33,9 +34,9 @@ export default function Team() {
         description="Lorem ipsum dolor sit amet consectetur. Et vestibulum enim nunc ultrices. Donec blandit
           sollicitudin vitae integer mauris sed. Mattis duis id viverra suscipit morbi."
       />
-      <div className="py-10 px-20">
-        <h1 className="text-white text-2xl py-8 font-bold">Our Team</h1>
-        <p className="w-5/6">
+      <div className={styles.text}>
+        <div className={styles.subtitle}>Our Team</div>
+        <p className={styles.description}>
           Our dedicated team @ 4 Future Leaders of Tomorrow is a non-profit charitable organization
           committed in preventing and ending homelessness, hunger and disparity in underprivileged
           communities. Everyone deserves a chance for a better future!. We are reaching out by
@@ -43,7 +44,7 @@ export default function Team() {
           educational supplies, referrals, toys or even bus passes
         </p>
       </div>
-      <div className="grid grid-cols-4">
+      <div className={styles.membersContainer}>
         {members.map((member) => (
           <MemberInfo key={member._id} member={member} />
         ))}
