+++ frontend/src/components/BackgroundHeader.tsx
