+++ frontend/src/components/MemberInfo.module.css
@@ -1,0 +1,26 @@
+@import url("https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Roboto+Slab:wght@100..900&display=swap");
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.profile {
+  width: 248px;
+  height: 249px;
+  border-radius: 9999px;
+  overflow: hidden;
+  margin-bottom: 23px;
+}
+
+.name {
+  font: var(--font-small-subtitle);
+  font-size: 28px;
+  line-height: 150%; /* 42px */
+  letter-spacing: 0.56px;
+}
+
+.role {
+  font: var(--font-body);
+}
